Extract Trino View functionality
