+++ plugin/trino-iceberg/src/main/java/io/trino/plugin/iceberg/catalog/AbstractTrinoCatalog.java
@@ -17,8 +17,6 @@ import com.google.common.collect.ImmutableMap;
 import io.trino.filesystem.TrinoFileSystem;
 import io.trino.plugin.base.CatalogName;
 import io.trino.plugin.hive.HiveMetadata;
-import io.trino.plugin.hive.HiveViewNotSupportedException;
-import io.trino.plugin.hive.ViewReaderUtil;
 import io.trino.plugin.iceberg.ColumnIdentity;
 import io.trino.plugin.iceberg.IcebergMaterializedViewDefinition;
 import io.trino.plugin.iceberg.IcebergUtil;
@@ -48,15 +46,12 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
-import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Throwables.throwIfUnchecked;
 import static com.google.common.collect.ImmutableList.toImmutableList;
 import static io.trino.plugin.hive.HiveMetadata.STORAGE_TABLE;
 import static io.trino.plugin.hive.HiveMetadata.TABLE_COMMENT;
 import static io.trino.plugin.hive.ViewReaderUtil.ICEBERG_MATERIALIZED_VIEW_COMMENT;
 import static io.trino.plugin.hive.ViewReaderUtil.PRESTO_VIEW_FLAG;
-import static io.trino.plugin.hive.ViewReaderUtil.isHiveOrPrestoView;
-import static io.trino.plugin.hive.ViewReaderUtil.isPrestoView;
 import static io.trino.plugin.iceberg.IcebergErrorCode.ICEBERG_FILESYSTEM_ERROR;
 import static io.trino.plugin.iceberg.IcebergMaterializedViewAdditionalProperties.STORAGE_SCHEMA;
 import static io.trino.plugin.iceberg.IcebergMaterializedViewAdditionalProperties.getStorageSchema;
@@ -74,31 +69,24 @@ import static org.apache.iceberg.Transactions.createTableTransaction;
 public abstract class AbstractTrinoCatalog
         implements TrinoCatalog
 {
-    // Be compatible with views defined by the Hive connector, which can be useful under certain conditions.
+    public static final String TRINO_CREATED_BY_VALUE = "Trino Iceberg connector";
     protected static final String TRINO_CREATED_BY = HiveMetadata.TRINO_CREATED_BY;
-    protected static final String TRINO_CREATED_BY_VALUE = "Trino Iceberg connector";
-    protected static final String PRESTO_VIEW_COMMENT = HiveMetadata.PRESTO_VIEW_COMMENT;
-    protected static final String PRESTO_VERSION_NAME = HiveMetadata.PRESTO_VERSION_NAME;
     protected static final String PRESTO_QUERY_ID_NAME = HiveMetadata.PRESTO_QUERY_ID_NAME;
-    protected static final String PRESTO_VIEW_EXPANDED_TEXT_MARKER = HiveMetadata.PRESTO_VIEW_EXPANDED_TEXT_MARKER;
 
     private final CatalogName catalogName;
     private final TypeManager typeManager;
     protected final IcebergTableOperationsProvider tableOperationsProvider;
-    private final String trinoVersion;
     private final boolean useUniqueTableLocation;
 
     protected AbstractTrinoCatalog(
             CatalogName catalogName,
             TypeManager typeManager,
             IcebergTableOperationsProvider tableOperationsProvider,
-            String trinoVersion,
             boolean useUniqueTableLocation)
     {
         this.catalogName = requireNonNull(catalogName, "catalogName is null");
         this.typeManager = requireNonNull(typeManager, "typeManager is null");
         this.tableOperationsProvider = requireNonNull(tableOperationsProvider, "tableOperationsProvider is null");
-        this.trinoVersion = requireNonNull(trinoVersion, "trinoVersion is null");
         this.useUniqueTableLocation = useUniqueTableLocation;
     }
 
@@ -197,56 +185,6 @@ public abstract class AbstractTrinoCatalog
         catch (IOException e) {
             throw new TrinoException(ICEBERG_FILESYSTEM_ERROR, format("Failed to delete directory %s of the table %s", tableLocation, schemaTableName), e);
         }
-    }
-
-    protected Optional<ConnectorViewDefinition> getView(
-            SchemaTableName viewName,
-            Optional<String> viewOriginalText,
-            String tableType,
-            Map<String, String> tableParameters,
-            Optional<String> tableOwner)
-    {
-        if (!isView(tableType, tableParameters)) {
-            // Filter out Tables and Materialized Views
-            return Optional.empty();
-        }
-
-        if (!isPrestoView(tableParameters)) {
-            // Hive views are not compatible
-            throw new HiveViewNotSupportedException(viewName);
-        }
-
-        checkArgument(viewOriginalText.isPresent(), "viewOriginalText must be present");
-        ConnectorViewDefinition definition = ViewReaderUtil.PrestoViewReader.decodeViewData(viewOriginalText.get());
-        // use owner from table metadata if it exists
-        if (tableOwner.isPresent() && !definition.isRunAsInvoker()) {
-            definition = new ConnectorViewDefinition(
-                    definition.getOriginalSql(),
-                    definition.getCatalog(),
-                    definition.getSchema(),
-                    definition.getColumns(),
-                    definition.getComment(),
-                    tableOwner,
-                    false);
-        }
-        return Optional.of(definition);
-    }
-
-    private static boolean isView(String tableType, Map<String, String> tableParameters)
-
-    {
-        return isHiveOrPrestoView(tableType) && PRESTO_VIEW_COMMENT.equals(tableParameters.get(TABLE_COMMENT));
-    }
-
-    protected Map<String, String> createViewProperties(ConnectorSession session)
-    {
-        return ImmutableMap.<String, String>builder()
-                .put(PRESTO_VIEW_FLAG, "true") // Ensures compatibility with views created by the Hive connector
-                .put(TRINO_CREATED_BY, TRINO_CREATED_BY_VALUE)
-                .put(PRESTO_VERSION_NAME, trinoVersion)
-                .put(PRESTO_QUERY_ID_NAME, session.getQueryId())
-                .put(TABLE_COMMENT, PRESTO_VIEW_COMMENT)
-                .buildOrThrow();
     }
 
     protected SchemaTableName createMaterializedViewStorageTable(ConnectorSession session, SchemaTableName viewName, ConnectorMaterializedViewDefinition definition)
