+++ plugin/trino-iceberg/src/main/java/io/trino/plugin/iceberg/catalog/hms/TrinoHiveCatalog.java
@@ -19,8 +19,8 @@ import io.airlift.log.Logger;
 import io.trino.filesystem.TrinoFileSystemFactory;
 import io.trino.plugin.base.CatalogName;
 import io.trino.plugin.hive.HiveSchemaProperties;
-import io.trino.plugin.hive.TableAlreadyExistsException;
-import io.trino.plugin.hive.ViewAlreadyExistsException;
+import io.trino.plugin.hive.TrinoViewHiveMetastore;
+import io.trino.plugin.hive.TrinoViewUtil;
 import io.trino.plugin.hive.metastore.Column;
 import io.trino.plugin.hive.metastore.Database;
 import io.trino.plugin.hive.metastore.HivePrincipal;
@@ -56,7 +56,6 @@ import java.util.Objects;
 import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 import static com.google.common.base.Preconditions.checkState;
 import static com.google.common.collect.ImmutableList.toImmutableList;
@@ -68,7 +67,6 @@ import static io.trino.plugin.hive.HiveType.HIVE_STRING;
 import static io.trino.plugin.hive.ViewReaderUtil.ICEBERG_MATERIALIZED_VIEW_COMMENT;
 import static io.trino.plugin.hive.ViewReaderUtil.encodeViewData;
 import static io.trino.plugin.hive.ViewReaderUtil.isHiveOrPrestoView;
-import static io.trino.plugin.hive.ViewReaderUtil.isPrestoView;
 import static io.trino.plugin.hive.ViewReaderUtil.isTrinoMaterializedView;
 import static io.trino.plugin.hive.metastore.MetastoreUtil.buildInitialPrivilegeSet;
 import static io.trino.plugin.hive.metastore.PrincipalPrivileges.NO_PRIVILEGES;
@@ -103,6 +101,7 @@ public class TrinoHiveCatalog
     public static final String DEPENDS_ON_TABLES = "dependsOnTables";
 
     private final CachingHiveMetastore metastore;
+    private final TrinoViewHiveMetastore trinoViewHiveMetastore;
     private final TrinoFileSystemFactory fileSystemFactory;
     private final boolean isUsingSystemSecurity;
     private final boolean deleteSchemaLocationsFallback;
@@ -112,16 +111,17 @@ public class TrinoHiveCatalog
     public TrinoHiveCatalog(
             CatalogName catalogName,
             CachingHiveMetastore metastore,
+            TrinoViewHiveMetastore trinoViewHiveMetastore,
             TrinoFileSystemFactory fileSystemFactory,
             TypeManager typeManager,
             IcebergTableOperationsProvider tableOperationsProvider,
-            String trinoVersion,
             boolean useUniqueTableLocation,
             boolean isUsingSystemSecurity,
             boolean deleteSchemaLocationsFallback)
     {
-        super(catalogName, typeManager, tableOperationsProvider, trinoVersion, useUniqueTableLocation);
+        super(catalogName, typeManager, tableOperationsProvider, useUniqueTableLocation);
         this.metastore = requireNonNull(metastore, "metastore is null");
+        this.trinoViewHiveMetastore = requireNonNull(trinoViewHiveMetastore, "viewsMetastoreHelper is null");
         this.fileSystemFactory = requireNonNull(fileSystemFactory, "fileSystemFactory is null");
         this.isUsingSystemSecurity = isUsingSystemSecurity;
         this.deleteSchemaLocationsFallback = deleteSchemaLocationsFallback;
@@ -319,7 +319,7 @@ public class TrinoHiveCatalog
         io.trino.plugin.hive.metastore.Table view = metastore.getTable(viewName.getSchemaName(), viewName.getTableName())
                 .orElseThrow(() -> new ViewNotFoundException(viewName));
 
-        ConnectorViewDefinition definition = getView(viewName, view.getViewOriginalText(), view.getTableType(), view.getParameters(), view.getOwner())
+        ConnectorViewDefinition definition = TrinoViewUtil.getView(viewName, view.getViewOriginalText(), view.getTableType(), view.getParameters(), view.getOwner())
                 .orElseThrow(() -> new ViewNotFoundException(viewName));
         ConnectorViewDefinition newDefinition = new ConnectorViewDefinition(
                 definition.getOriginalSql(),
@@ -339,7 +339,7 @@ public class TrinoHiveCatalog
         io.trino.plugin.hive.metastore.Table view = metastore.getTable(viewName.getSchemaName(), viewName.getTableName())
                 .orElseThrow(() -> new ViewNotFoundException(viewName));
 
-        ConnectorViewDefinition definition = getView(viewName, view.getViewOriginalText(), view.getTableType(), view.getParameters(), view.getOwner())
+        ConnectorViewDefinition definition = TrinoViewUtil.getView(viewName, view.getViewOriginalText(), view.getTableType(), view.getParameters(), view.getOwner())
                 .orElseThrow(() -> new ViewNotFoundException(viewName));
         ConnectorViewDefinition newDefinition = new ConnectorViewDefinition(
                 definition.getOriginalSql(),
@@ -392,43 +392,7 @@ public class TrinoHiveCatalog
     @Override
     public void createView(ConnectorSession session, SchemaTableName schemaViewName, ConnectorViewDefinition definition, boolean replace)
     {
-        if (isUsingSystemSecurity) {
-            definition = definition.withoutOwner();
-        }
-
-        io.trino.plugin.hive.metastore.Table.Builder tableBuilder = io.trino.plugin.hive.metastore.Table.builder()
-                .setDatabaseName(schemaViewName.getSchemaName())
-                .setTableName(schemaViewName.getTableName())
-                .setOwner(isUsingSystemSecurity ? Optional.empty() : Optional.of(session.getUser()))
-                .setTableType(org.apache.hadoop.hive.metastore.TableType.VIRTUAL_VIEW.name())
-                .setDataColumns(ImmutableList.of(new Column("dummy", HIVE_STRING, Optional.empty())))
-                .setPartitionColumns(ImmutableList.of())
-                .setParameters(createViewProperties(session))
-                .setViewOriginalText(Optional.of(encodeViewData(definition)))
-                .setViewExpandedText(Optional.of(PRESTO_VIEW_EXPANDED_TEXT_MARKER));
-
-        tableBuilder.getStorageBuilder()
-                .setStorageFormat(VIEW_STORAGE_FORMAT)
-                .setLocation("");
-        io.trino.plugin.hive.metastore.Table table = tableBuilder.build();
-        PrincipalPrivileges principalPrivileges = isUsingSystemSecurity ? NO_PRIVILEGES : buildInitialPrivilegeSet(session.getUser());
-
-        Optional<io.trino.plugin.hive.metastore.Table> existing = metastore.getTable(schemaViewName.getSchemaName(), schemaViewName.getTableName());
-        if (existing.isPresent()) {
-            if (!replace || !isPrestoView(existing.get())) {
-                throw new ViewAlreadyExistsException(schemaViewName);
-            }
-
-            metastore.replaceTable(schemaViewName.getSchemaName(), schemaViewName.getTableName(), table, principalPrivileges);
-            return;
-        }
-
-        try {
-            metastore.createTable(table, principalPrivileges);
-        }
-        catch (TableAlreadyExistsException e) {
-            throw new ViewAlreadyExistsException(e.getTableName());
-        }
+        trinoViewHiveMetastore.createView(session, schemaViewName, definition, replace);
     }
 
     @Override
@@ -448,46 +412,19 @@ public class TrinoHiveCatalog
     @Override
     public void dropView(ConnectorSession session, SchemaTableName schemaViewName)
     {
-        if (getView(session, schemaViewName).isEmpty()) {
-            throw new ViewNotFoundException(schemaViewName);
-        }
-
-        try {
-            metastore.dropTable(schemaViewName.getSchemaName(), schemaViewName.getTableName(), true);
-        }
-        catch (TableNotFoundException e) {
-            throw new ViewNotFoundException(e.getTableName());
-        }
+        trinoViewHiveMetastore.dropView(schemaViewName);
     }
 
     @Override
     public List<SchemaTableName> listViews(ConnectorSession session, Optional<String> namespace)
     {
-        return listNamespaces(session, namespace).stream()
-                .flatMap(this::listViews)
-                .collect(toImmutableList());
-    }
-
-    private Stream<SchemaTableName> listViews(String schema)
-    {
-        // Filter on PRESTO_VIEW_COMMENT to distinguish from materialized views
-        return metastore.getTablesWithParameter(schema, TABLE_COMMENT, PRESTO_VIEW_COMMENT).stream()
-                .map(table -> new SchemaTableName(schema, table));
+        return trinoViewHiveMetastore.listViews(namespace);
     }
 
     @Override
     public Optional<ConnectorViewDefinition> getView(ConnectorSession session, SchemaTableName viewName)
     {
-        if (isHiveSystemSchema(viewName.getSchemaName())) {
-            return Optional.empty();
-        }
-        return metastore.getTable(viewName.getSchemaName(), viewName.getTableName())
-                .flatMap(view -> getView(
-                        viewName,
-                        view.getViewOriginalText(),
-                        view.getTableType(),
-                        view.getParameters(),
-                        view.getOwner()));
+        return trinoViewHiveMetastore.getView(viewName);
     }
 
     @Override
