+++ plugin/trino-iceberg/src/main/java/io/trino/plugin/iceberg/catalog/hms/TrinoHiveCatalogFactory.java
@@ -16,7 +16,9 @@ package io.trino.plugin.iceberg.catalog.hms;
 import io.trino.filesystem.TrinoFileSystemFactory;
 import io.trino.plugin.base.CatalogName;
 import io.trino.plugin.hive.NodeVersion;
+import io.trino.plugin.hive.TrinoViewHiveMetastore;
 import io.trino.plugin.hive.metastore.HiveMetastoreFactory;
+import io.trino.plugin.hive.metastore.cache.CachingHiveMetastore;
 import io.trino.plugin.iceberg.IcebergConfig;
 import io.trino.plugin.iceberg.IcebergSecurityConfig;
 import io.trino.plugin.iceberg.catalog.IcebergTableOperationsProvider;
@@ -31,6 +33,7 @@ import java.util.Optional;
 
 import static io.trino.plugin.hive.metastore.cache.CachingHiveMetastore.memoizeMetastore;
 import static io.trino.plugin.iceberg.IcebergSecurityConfig.IcebergSecurity.SYSTEM;
+import static io.trino.plugin.iceberg.catalog.AbstractTrinoCatalog.TRINO_CREATED_BY_VALUE;
 import static java.util.Objects.requireNonNull;
 
 public class TrinoHiveCatalogFactory
@@ -71,13 +74,14 @@ public class TrinoHiveCatalogFactory
     @Override
     public TrinoCatalog create(ConnectorIdentity identity)
     {
+        CachingHiveMetastore metastore = memoizeMetastore(metastoreFactory.createMetastore(Optional.of(identity)), 1000);
         return new TrinoHiveCatalog(
                 catalogName,
-                memoizeMetastore(metastoreFactory.createMetastore(Optional.of(identity)), 1000),
+                metastore,
+                new TrinoViewHiveMetastore(isUsingSystemSecurity, metastore, trinoVersion, TRINO_CREATED_BY_VALUE),
                 fileSystemFactory,
                 typeManager,
                 tableOperationsProvider,
-                trinoVersion,
                 isUniqueTableLocation,
                 isUsingSystemSecurity,
                 deleteSchemaLocationsFallback);
